Deletes nested folders as well.
I had a folder named "Markdown Monster Weblog Posts" in the previous "West Wind Markdown Monster". The folder delete operation without recursive delete could not delete it and was throwing an exception.

For me this fixed #35 as well, but I'm not sure if this is an acceptable fix for the issue. It might be still nice to ask for a new license key when the current one is invalid somehow.